+++ src/design/plone/contenttypes/behaviors/show_modified.py
@@ -1,8 +1,6 @@
 # -*- coding: utf-8 -*-
 from design.plone.contenttypes import _
-from design.plone.contenttypes.controlpanels.settings import (
-    IDesignPloneSettings,
-)
+from design.plone.contenttypes.controlpanels.settings import IDesignPloneSettings
 from plone import api
 from plone.autoform.interfaces import IFormFieldProvider
 from plone.dexterity.interfaces import IDexterityContent
@@ -12,7 +10,7 @@ from zope.component import adapter
 from zope.interface import provider, implementer
 
 
-def showModifiedDefaultValue():
+def showModifiedDefaultValue(context=None):
     return api.portal.get_registry_record(
         "show_modified_default", interface=IDesignPloneSettings, default=False
     )
@@ -22,9 +20,7 @@ def showModifiedDefaultValue():
 class IShowModified(model.Schema):
 
     show_modified = schema.Bool(
-        title=_(
-            "show_modified_label", default="Mostra la data di ultima modifica"
-        ),
+        title=_("show_modified_label", default="Mostra la data di ultima modifica"),
         description=_(
             "show_modified_help",
             default="Se attivo, verrà mostrata la data di ultima modifica in "